More clang-tidy fixes (#64)
* uint8_t underlying type for enum

* Changing the argument int nx to const int in markWallBoundary()

* Removed redundant boolean literal

* Changing void return type to void const

* Changing clang version from 17 to 19 for CI

* Updating .clang-tidy to eliminate outdated option

* Replacing macros with enum definitions

* Changing void const to void

* Removing redundant explicit casting

* Addressing the readability-math-missing-parentheses warning

* More readability-math-missing-parentheses fixes

* static_cast<ptrdiff_t>

* static_cast<unsigned long>

* Removing inline specifiers from implicitly inlined functions to eliminate the warning readability-redundant-inline-specifier

* static_cast<ptrdiff_t>() instances in MeshBlock.C

* Changing enum to define in tioga.h

* Fixing enum declarations

+++ src/exchangeDonors.C
@@ -139,7 +139,7 @@ void tioga::exchangeDonors()
 
     for (int n = 0; n < nblocks; n++) {
         for (int i = 0; i < nrecords[n]; i++) {
-            int const k = donorRecords[n][static_cast<int>(3 * i)];
+            int const k = donorRecords[n][3 * i];
             sndPack[k].nints += 2;
             sndPack[k].nreals++;
         }
@@ -153,9 +153,9 @@ void tioga::exchangeDonors()
 
     for (int n = 0; n < nblocks; n++) {
         for (int i = 0; i < nrecords[n]; i++) {
-            int const k = donorRecords[n][static_cast<int>(3 * i)];
-            sndPack[k].intData[ixOffset[k]++] = donorRecords[n][3 * i + 1];
-            sndPack[k].intData[ixOffset[k]++] = donorRecords[n][3 * i + 2];
+            int const k = donorRecords[n][3 * i];
+            sndPack[k].intData[ixOffset[k]++] = donorRecords[n][(3 * i) + 1];
+            sndPack[k].intData[ixOffset[k]++] = donorRecords[n][(3 * i) + 2];
             sndPack[k].realData[rxOffset[k]++] = receptorResolution[n][i];
         }
     }
@@ -213,7 +213,7 @@ void tioga::exchangeDonors()
 
     for (int n = 0; n < nblocks; n++) {
         for (int i = 0; i < nrecords[n]; i++) {
-            int const k = donorRecords[n][static_cast<int>(3 * i)];
+            int const k = donorRecords[n][3 * i];
             sndPack[k].nints += 2;
         }
     }
@@ -223,9 +223,9 @@ void tioga::exchangeDonors()
 
     for (int n = 0; n < nblocks; n++) {
         for (int i = 0; i < nrecords[n]; i++) {
-            int const k = donorRecords[n][static_cast<int>(3 * i)];
-            sndPack[k].intData[ixOffset[k]++] = donorRecords[n][3 * i + 1];
-            sndPack[k].intData[ixOffset[k]++] = donorRecords[n][3 * i + 2];
+            int const k = donorRecords[n][3 * i];
+            sndPack[k].intData[ixOffset[k]++] = donorRecords[n][(3 * i) + 1];
+            sndPack[k].intData[ixOffset[k]++] = donorRecords[n][(3 * i) + 2];
         }
     }
     //
@@ -264,7 +264,7 @@ void tioga::exchangeDonors()
     std::fill(ixOffset.begin(), ixOffset.end(), 0);
     for (int n = 0; n < nblocks; n++) {
         for (int i = 0; i < nrecords[n]; i++) {
-            int const k = donorRecords[n][static_cast<int>(3 * i)];
+            int const k = donorRecords[n][3 * i];
             sndPack[k].nints += 2;
         }
     }
@@ -273,9 +273,9 @@ void tioga::exchangeDonors()
     }
     for (int n = 0; n < nblocks; n++) {
         for (int i = 0; i < nrecords[n]; i++) {
-            int const k = donorRecords[n][static_cast<int>(3 * i)];
-            sndPack[k].intData[ixOffset[k]++] = donorRecords[n][3 * i + 1];
-            sndPack[k].intData[ixOffset[k]++] = donorRecords[n][3 * i + 2];
+            int const k = donorRecords[n][3 * i];
+            sndPack[k].intData[ixOffset[k]++] = donorRecords[n][(3 * i) + 1];
+            sndPack[k].intData[ixOffset[k]++] = donorRecords[n][(3 * i) + 2];
         }
     }
     //
